Finally added the leaderboards.html, hasn't changed much but it needs to be here
Have a matches played, won and the last 20 matches played. I haven't
tested the last 20  stuff at all, but it is pretty straight forward so I
hope I wrote it properly.

I think because it's getting (ha!) late I'm not able to  think properly,
meaning I'm starting to think I stand a chance. Going to play around with
some Python stuff, see what I can muster in about an hour or so, including
PEP8'ing the heck out of it, and then working on exporting it for
FluidWare dudes to use.